Refactoring RL howtos
