+++ src/mlpro/rl/examples/howto_rl_003_train_agent_with_own_policy_on_gym_environment.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-003_Train_agent_with_own_policy_on_gym_environment.py
+## -- Module  : howto_rl_003_train_agent_with_own_policy_on_gym_environment.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
