+++ src/mlpro/rl/examples/howto_rl_006_run_own_agents_with_petting_zoo_environment.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-006_Run_own_agents_with_petting_zoo_environment.py
+## -- Module  : howto_rl_006_run_own_agents_with_petting_zoo_environment.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
