+++ src/mlpro/rl/examples/howto_rl_007_train_wrapped_SB3_policy.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-007_Train_wrapped_SB3_policy.py
+## -- Module  : howto_rl_007_train_wrapped_SB3_policy.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
