+++ src/mlpro/rl/examples/howto_rl_008_wrap_mlpro_environment_to_gym_environment.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-008_Wrap_MLPro_environment_to_Gym_environment.py 
+## -- Module  : howto_rl_008_wrap_mlpro_environment_to_gym_environment.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
