+++ src/mlpro/rl/examples/howto_rl_009_wrap_mlpro_environment_to_pettingzoo_environment.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-009_Wrap_MLPro_environment_to_PettingZoo_environment.py
+## -- Module  : howto_rl_009_wrap_mlpro_environment_to_pettingzoo_environment.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
