+++ src/mlpro/rl/examples/howto_rl_013_model_based_reinforcement_learning.py
@@ -1,44 +1,59 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-017_Train_wrapped_SB3_policy_on_MultiGeo_environment.py
+## -- Module  : howto_rl_013_model_based_reinforcement_learning.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
-## -- 2021-12-19  0.0.0     MRD      Creation
-## -- 2021-12-19  1.0.0     MRD      Initial Release
-## -- 2021-12-23  1.0.1     DA       Minor fix 
+## -- 2021-12-17  0.0.0     MRD       Creation
+## -- 2021-12-17  1.0.0     MRD       Released first version
+## -- 2022-01-01  1.0.1     MRD       Refactoring due to new model implementation
+## -- 2022-05-20  1.0.2     MRD       Add HTMEnvModel
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.1 (2021-12-23)
+Ver. 1.0.2 (2022-05-20)
 
-This module shows how to use SB3 wrapper to train Multi Geometry Robot.
+This module demonstrates model-based reinforcement learning (MBRL).
 """
 
-from mlpro.bf.math import *
+
+import torch
+
+from mlpro.bf.ml import *
 from mlpro.rl.models import *
-from mlpro.rl.pool.envs.multigeorobot import MultiGeo
+from mlpro.rl.pool.envs.robotinhtm import RobotHTM
 from stable_baselines3 import PPO
 from mlpro.wrappers.sb3 import WrPolicySB32MLPro
+from mlpro.rl.pool.envmodels.mlp_robotinhtm import MLPEnvModel
+from mlpro.rl.pool.envmodels.htm_robotinhtm import HTMEnvModel
+
 from pathlib import Path
 
 
-# 1 Implement your own RL scenario
-class ScenarioMultiGeoPPO(RLScenario):
-    C_NAME = 'Matrix'
+class ActualTraining(RLTraining):
+    C_NAME = "Actual"
+
+
+# Implement RL Scenario for the actual environment to train the environment model
+class ScenarioRobotHTMActual(RLScenario):
+    C_NAME = "Matrix1"
 
     def _setup(self, p_mode, p_ada, p_logging):
         # 1 Setup environment
-        self._env = MultiGeo(p_logging=p_logging)
+        self._env = RobotHTM(p_logging=True)
+
+        policy_kwargs = dict(activation_fn=torch.nn.Tanh,
+                             net_arch=[dict(pi=[128, 128], vf=[128, 128])])
 
         policy_sb3 = PPO(
             policy="MlpPolicy",
-            n_steps=20,
+            n_steps=100,
             env=None,
             _init_setup_model=False,
-            device="cpu",
-            seed=1)
+            policy_kwargs=policy_kwargs,
+            device="cpu"
+        )
 
         policy_wrapped = WrPolicySB32MLPro(
             p_sb3_policy=policy_sb3,
@@ -46,32 +61,59 @@ class ScenarioMultiGeoPPO(RLScenario):
             p_observation_space=self._env.get_state_space(),
             p_action_space=self._env.get_action_space(),
             p_ada=p_ada,
-            p_logging=p_logging)
+            p_logging=p_logging,
+        )
+
+        mb_training_param = dict(p_cycle_limit=100,
+                                 p_cycles_per_epi_limit=100,
+                                 p_max_stagnations=0,
+                                 p_collect_states=False,
+                                 p_collect_actions=False,
+                                 p_collect_rewards=False,
+                                 p_collect_training=False)
 
         # 2 Setup standard single-agent with own policy
         return Agent(
             p_policy=policy_wrapped,
-            p_envmodel=None,
-            p_name='Smith',
+            p_envmodel=HTMEnvModel(),
+            p_em_mat_thsld=0.5,
+            p_name="Smith1",
             p_ada=p_ada,
-            p_logging=p_logging
+            p_logging=p_logging,
+            **mb_training_param
         )
 
 
 # 3 Train agent in scenario
 now = datetime.now()
 
-training = RLTraining(
-    p_scenario_cls=ScenarioMultiGeoPPO,
-    p_cycle_limit=1000,
-    p_cycles_per_epi_limit=-1,
+if __name__ == "__main__":
+    # 3.1 Parameters for demo mode
+    cycle_limit = 300000
+    logging = Log.C_LOG_ALL
+    visualize = True
+    path = str(Path.home())
+    plotting = True
+
+else:
+    # 3.2 Parameters for internal unit test
+    cycle_limit = 100
+    logging = Log.C_LOG_NOTHING
+    visualize = False
+    path = None
+    plotting = False
+
+training = ActualTraining(
+    p_scenario_cls=ScenarioRobotHTMActual,
+    p_cycle_limit=cycle_limit,
+    p_cycles_per_epi_limit=100,
     p_collect_states=True,
     p_collect_actions=True,
     p_collect_rewards=True,
     p_collect_training=True,
-    p_visualize=False,
-    p_path=str(Path.home()),
-    p_logging=Log.C_LOG_ALL)
+    p_path=path,
+    p_logging=logging,
+)
 
 training.run()
 
@@ -115,12 +157,14 @@ class MyDataPlotting(DataPlotting):
 
 
 # 5 Plotting 1 MLpro
-data_printing = {"Cycle": [False],
-                 "Day": [False],
-                 "Second": [False],
-                 "Microsecond": [False],
-                 "Smith": [True, -1]}
+data_printing = {
+    "Cycle": [False],
+    "Day": [False],
+    "Second": [False],
+    "Microsecond": [False],
+    "Smith1": [True, -1],
+}
 
 mem = training.get_results().ds_rewards
-mem_plot = MyDataPlotting(mem, p_showing=True, p_printing=data_printing)
+mem_plot = MyDataPlotting(mem, p_showing=plotting, p_printing=data_printing)
 mem_plot.get_plots()
