+++ src/mlpro/rl/examples/howto_rl_015_train_wrapped_sb3_policy_with_stagnation_detection.py
@@ -1,40 +1,40 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-RL-018_Train_and_reload_single_agent.py
+## -- Module  : howto_rl_015_train_wrapped_sb3_policy_with_stagnation_detection.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
-## -- 2022-01-28  0.0.0     MRD      Creation
-## -- 2022-01-28  1.0.0     MRD      Released first version
-## -- 2022-05-19  1.0.1     MRD      Re-use the agent not for the re-training process
-## --                                Remove commenting and numbering
-## -- 2022-05-19  1.0.2     MRD      Re-add the commneting and reformat the numbering in comment
+## -- 2022-01-20  0.0.0     MRD      Creation
+## -- 2022-01-20  1.0.0     MRD      Released first version
+## -- 2022-05-17  1.0.1     DA       Just a litte comment maintenance
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.2 (2022-05-19)
+Ver. 1.0.1 (2022-05-17)
 
-This module shows how to train a single agent and load it again to do some extra cycles
+This module shows how to train with SB3 Wrapper and stagnation detection
 """
 
 import gym
-from stable_baselines3 import PPO
+from stable_baselines3 import A2C, PPO, DQN, DDPG, SAC
 from mlpro.rl.models import *
 from mlpro.wrappers.openai_gym import WrEnvGYM2MLPro
 from mlpro.wrappers.sb3 import WrPolicySB32MLPro
 from pathlib import Path
+
+
 
 # 1 Implement your own RL scenario
 class MyScenario(RLScenario):
     C_NAME = 'Matrix'
 
     def _setup(self, p_mode, p_ada, p_logging):
-        # 1.1 Setup environment
+        # 1 Setup environment
         gym_env = gym.make('CartPole-v1')
         self._env = WrEnvGYM2MLPro(gym_env, p_logging=p_logging)
 
-        # 1.2 Setup Policy From SB3
+        # 2 Instantiate PPO Policy from SB3
         policy_sb3 = PPO(
             policy="MlpPolicy",
             n_steps=5,
@@ -43,7 +43,7 @@ class MyScenario(RLScenario):
             device="cpu",
             seed=1)
 
-        # 1.3 Wrap the policy
+        # 3 Wrap the policy
         policy_wrapped = WrPolicySB32MLPro(
             p_sb3_policy=policy_sb3,
             p_cycle_limit=self._cycle_limit,
@@ -52,7 +52,7 @@ class MyScenario(RLScenario):
             p_ada=p_ada,
             p_logging=p_logging)
 
-        # 1.4 Setup standard single-agent with own policy
+        # 4 Setup standard single-agent with own policy
         return Agent(
             p_policy=policy_wrapped,
             p_envmodel=None,
@@ -61,8 +61,12 @@ class MyScenario(RLScenario):
             p_logging=p_logging
         )
 
+
+
+# 2 Create scenario and start training
+
 if __name__ == "__main__":
-    # Parameters for demo mode
+    # 2.1 Parameters for demo mode
     cycle_limit = 5000
     adaptation_limit = 50
     stagnation_limit = 5
@@ -73,7 +77,7 @@ if __name__ == "__main__":
     path = str(Path.home())
 
 else:
-    # Parameters for internal unit test
+    # 2.2 Parameters for internal unit test
     cycle_limit = 50
     adaptation_limit = 5
     stagnation_limit = 5
@@ -83,7 +87,7 @@ else:
     visualize = False
     path = None
 
-# 2 Create scenario and start training
+# 2.3 Create and run training object
 training = RLTraining(
     p_scenario_cls=MyScenario,
     p_cycle_limit=cycle_limit,
@@ -95,35 +99,4 @@ training = RLTraining(
     p_visualize=visualize,
     p_logging=logging)
 
-# 3 Create scenario and start training
 training.run()
-
-# 4 Save the training path for loading the agent model file
-training_path = training._root_path
-
-# We start from the beginning, in this case we load an existing model
-# 5 Implement your own RL scenario with an existing model
-class MyNdScenario(RLScenario):
-    C_NAME = 'Matrix2'
-
-    def _setup(self, p_mode, p_ada, p_logging):
-        # 5.1 Setup environment
-        gym_env = gym.make('CartPole-v1')
-        self._env = WrEnvGYM2MLPro(gym_env, p_logging=p_logging)
-
-        # 5.2 In this example we use previous training from the same file
-        # To make easier, we retrieve the save path from the previous training
-        return self.load(training_path, "trained model.pkl")
-
-# 6 Instatiate new scenario
-scenario = MyNdScenario(p_mode=Mode.C_MODE_SIM, 
-                        p_ada=False,
-                        p_cycle_limit=cycle_limit,
-                        p_visualize=visualize,
-                        p_logging=logging)
-
-# 7 Reset Scenario
-scenario.reset()  
-
-# 8 Run Scenario
-scenario.run()
